# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test task API Documentation",
        "description": "Test task API Documentation Description",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "Local API server"
        }
    ],
    "paths": {
...
        "/api/brands": {
            "get": {
                "tags": [
                    "Brands Routes"
                ],
                "summary": "Brands list",
                "description": "Brands list.",
                "operationId": "brandList",
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Brands Routes"
                ],
                "summary": "Brand add",
                "description": "Brand add.",
                "operationId": "BrandCreate",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Brand name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Brand Name"
                    },
                    {
                        "name": "brand_id",
                        "in": "query",
                        "description": "Brand Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "district_id",
                        "in": "query",
                        "description": "District Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "12"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized!"
                    },
                    "422": {
                        "description": "Validation error!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/brands/{id}": {
            "get": {
                "tags": [
                    "Brands Routes"
                ],
                "summary": "Brand show",
                "description": "Brand show.",
                "operationId": "BrandShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Brand id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "2"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Brand not found!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Brands Routes"
                ],
                "summary": "Brand update",
                "description": "Brand update.",
                "operationId": "BrandUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Brand id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "143"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Brand name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "New Brand Name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Brands Routes"
                ],
                "summary": "Brand delete",
                "description": "Brand delete info.",
                "operationId": "BrandDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Brand id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "143"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Brand not found!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Delete error!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized!"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}